requirements.txt
PyJWT>=2.0.0
bcrypt>=4.0.0
